Refactor/home components 분리
