+++ app/(nav)/home/page.jsx
@@ -1,7 +1,7 @@
 "use client";
 
-import HomeHeader from './home-header';
-import HomeBody1 from './home-body1';
+import HomeHeader from '../../../components/home/home-header';
+import HomeBody1 from '../../../components/home/home-body1';
 import HomeBody2 from './home-body2';
 
 const Home = () => {
